requirements.txt
asgiref==3.4.1
backcall==0.2.0
decorator==5.1.0
dj-database-url==0.5.0
Django==3.2.8
django-safedelete==1.1.0
gunicorn==20.1.0
ipdb==0.13.9
ipython>=7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
ptyprocess==0.7.0
Pygments==2.10.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
traitlets==4.3.3
typing-extensions==3.10.0.2
wcwidth==0.2.5
